requirement.txt
alabaster==0.7.10
alembic==0.9.10
amqp==2.3.2
anaconda-client==1.6.5
anaconda-navigator==1.6.9
anaconda-project==0.8.0
aniso8601==3.0.2
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.22.0
astroid==1.5.3
astropy==2.0.2
attrs==17.4.0
Automat==0.6.0
awesome-slugify==1.6.5
Babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
billiard==3.5.0.4
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.0.0
blinker==1.4
bokeh==0.12.10
boto==2.48.0
Bottleneck==1.2.1
celery==4.2.1
certifi==2018.4.16
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
colorama==0.3.9
conda==4.5.4
conda-build==3.0.27
conda-verify==2.0.0
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.0.3
cssselect==1.0.3
cycler==0.10.0
cymysql==0.9.12
Cython==0.26.1
cytoolz==0.8.2
dask==0.15.3
datashape==0.5.4
decorator==4.1.2
diff-match-patch==20121119
distributed==1.19.1
Django==1.10.6
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-password-reset==1.0
django-ranged-response==0.2.0
django-reversion==3.0.0
django-simple-captcha==0.4.6
djangorestframework==3.8.2
docutils==0.14
dominate==2.3.1
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.12
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.3
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
flask-paginate==0.5.1
Flask-Redis==0.3.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.15.2
gevent==1.2.2
glob2==0.5
gmpy2==2.0.8
greenlet==0.4.13
h5py==2.7.0
heapdict==1.0.0
html5lib==0.999999999
httplib2==0.9.2
hyperlink==18.0.0
idna==2.6
imageio==2.2.0
imagesize==0.7.1
incremental==17.5.0
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
kombu==4.2.1
lazy-object-proxy==1.3.1
llvmlite==0.20.0
locket==0.2.0
lxml==4.1.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
mistune==0.7.4
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
mysql-connector-python==8.0.12
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numba==0.35.0+6.gaa35fb1
numexpr==2.6.2
numpy==1.13.3
numpydoc==0.7.0
odfpy==1.3.6
odo==0.5.1
olefile==0.44
openpyxl==2.4.8
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
parsel==1.4.0
partd==0.3.8
path.py==10.3.1
pathlib2==2.3.0
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
pkginfo==1.4.1
ply==3.10
prompt-toolkit==1.0.15
protobuf==3.6.0
psutil==5.4.0
ptyprocess==0.5.2
py==1.4.34
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
PyDispatcher==2.0.5
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.7.4
pymongo==3.7.1
PyMySQL==0.9.2
pyodbc==4.0.17
pyOpenSSL==17.2.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.2.1
python-dateutil==2.6.1
python-editor==1.0.3
python-memcached==1.59
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
qiniu==7.2.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
queuelib==1.5.0
redis==2.10.6
regex==2018.7.11
requests==2.18.4
rope==0.10.5
ruamel-yaml==0.11.14
scikit-image==0.13.0
scikit-learn==0.19.1
scipy==0.19.1
Scrapy==1.5.0
seaborn==0.8
service-identity==17.0.0
shortuuid==0.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
slugify==0.0.1
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.4
SQLAlchemy==1.1.13
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tablib==0.12.1
tblib==1.3.2
terminado==0.6
testpath==0.3.1
toolz==0.8.2
tornado==4.5.2
traitlets==4.3.2
Twisted==17.9.0
typing==3.6.2
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
visitor==0.1.3
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.2
wrapt==1.10.11
WTForms==2.2.1
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.4
xlwt==1.2.0
yunpian-python-sdk==1.0.0
zict==0.1.3
zope.interface==4.4.3
